newspaper_collector/requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==3.7.3
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.3.2
mysql-connector-python==2.0.4
newspaper3k==0.2.8
nltk==3.4
numpy==1.16.2
pandas==0.24.1
Pillow==5.4.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-file==1.4.3
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
SQLAlchemy==1.3.1
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.1
